app/requirements.txt
flask
torch
torchvision
Pillow
numpy
facenet-pytorch
